PeggleDeluxe1.01/requirements.txt
pypiwin32
pillow
python-dotenv
twitchAPI